Refactoring Signature class in typescript
Unsure what the correct convention is when defining types like the VRS object, if it should be in the class file itself or in the types file - I saw it both ways when looking through your and 0x's TypeScript code. I tried to leave the logic and actual code as untouched as possible, although I did catch a bug with how you were addressing the 0/1->27/28 vrs thing. Looks like the extra fromRpcSig call upon fail fixes this, but I think this block could be removed now.